+++ project/Dependencies.scala
@@ -8,7 +8,7 @@ object Dependencies {
   val scalaLoggingV = "3.7.2"
   val scalaTestV    = "3.1.1"
   val circeVersion = "0.13.0"
-  val http4sVersion = "0.21.0"
+  val http4sVersion = "0.21.4"
 
   def excludeGuavaJDK5(m: ModuleID): ModuleID = m.exclude("com.google.guava", "guava-jdk5")
 
@@ -31,7 +31,7 @@ object Dependencies {
 
   val selenium: ModuleID = "org.seleniumhq.selenium" % "selenium-java" % "3.11.0" % "test"
 
-  val catsEffect: ModuleID = "org.typelevel" %% "cats-effect" % "2.1.2"
+  val catsEffect: ModuleID = "org.typelevel" %% "cats-effect" % "2.1.3"
 
   // metrics-scala transitively pulls in io.dropwizard.metrics:metrics-core
   val metricsScala: ModuleID =      "nl.grons"              %% "metrics-scala"    % "3.5.6"
@@ -57,13 +57,13 @@ object Dependencies {
 
   val googleRpc2: ModuleID =               "io.grpc" % "grpc-core" % "1.28.0"
   val googleFirestore: ModuleID = "com.google.cloud" % "google-cloud-firestore" % "0.71.0-beta"
-  val googleStorageNew: ModuleID = "com.google.cloud" % "google-cloud-storage" % "1.100.0"
+  val googleStorageNew: ModuleID = "com.google.cloud" % "google-cloud-storage" % "1.107.0"
   val googleStorageLocal: ModuleID = "com.google.cloud" % "google-cloud-nio" % "0.111.0-alpha" % "test"
-  val googlePubsubNew: ModuleID = "com.google.cloud" % "google-cloud-pubsub" % "1.62.0"
+  val googlePubsubNew: ModuleID = "com.google.cloud" % "google-cloud-pubsub" % "1.105.0"
   val googleKms: ModuleID = "com.google.cloud" % "google-cloud-kms" % "0.77.0-beta"
-  val googleDataproc: ModuleID =    "com.google.cloud" % "google-cloud-dataproc" % "0.111.0"
-  val googleComputeNew: ModuleID = "com.google.cloud" % "google-cloud-compute" % "0.117.0-alpha"
-  val googleContainer: ModuleID = "com.google.cloud" % "google-cloud-container" % "1.0.0"
+  val googleDataproc: ModuleID =    "com.google.cloud" % "google-cloud-dataproc" % "0.122.1"
+  val googleComputeNew: ModuleID = "com.google.cloud" % "google-cloud-compute" % "0.118.0-alpha"
+  val googleContainer: ModuleID = "com.google.cloud" % "google-cloud-container" % "1.0.4"
   val kubernetesClient: ModuleID = "io.kubernetes" % "client-java" % "5.0.0" % "compile"
   //the below v1 module is a dependency for v2 because it contains the OAuth scopes necessary to created scoped credentials
   val googleContainerV1: ModuleID = "com.google.apis" % "google-api-services-container" % s"v1-rev74-$googleV"
@@ -73,15 +73,14 @@ object Dependencies {
   val circeParser: ModuleID = "io.circe" %% "circe-parser" % circeVersion
   val circeGeneric: ModuleID = "io.circe" %% "circe-generic" % circeVersion % "test"
   val circeFs2: ModuleID = "io.circe" %% "circe-fs2" % circeVersion
-  val cats: ModuleID = "org.typelevel" %% "cats-core" % "2.0.0"
-  val log4cats = "io.chrisdavenport" %% "log4cats-slf4j"   % "1.0.0"
-  val catsMtl = "org.typelevel" %% "cats-mtl-core" % "0.7.0"
+  val log4cats = "io.chrisdavenport" %% "log4cats-slf4j"   % "1.0.1"
+  val catsMtl = "org.typelevel" %% "cats-mtl-core" % "0.7.1"
 
   val http4sCirce = "org.http4s" %% "http4s-circe" % http4sVersion
   val http4sBlazeClient = "org.http4s" %% "http4s-blaze-client" % http4sVersion
   val http4sDsl = "org.http4s"      %% "http4s-dsl"          % http4sVersion
 
-  val fs2Io: ModuleID = "co.fs2" %% "fs2-io" % "2.0.1"
+  val fs2Io: ModuleID = "co.fs2" %% "fs2-io" % "2.3.0"
   val rawlsModel: ModuleID = "org.broadinstitute.dsde" %% "rawls-model" % "0.1-0d02c8ce-SNAP" exclude("com.typesafe.scala-logging", "scala-logging_2.11") exclude("com.typesafe.akka", "akka-stream_2.11")
   val newRelic: ModuleID = "com.newrelic.agent.java" % "newrelic-api" % "5.0.0"
   val openCensusApi: ModuleID = "io.opencensus" % "opencensus-api" % "0.26.0"
@@ -106,7 +105,7 @@ object Dependencies {
     akkaTestkit,
     mockito,
     scalaTestMockito,
-    cats
+    "org.typelevel" %% "cats-core" % "2.0.0" //This is the last version supports scala 2.11, which is use in orchestration
   )
 
   val modelDependencies = commonDependencies ++ Seq(
@@ -169,7 +168,8 @@ object Dependencies {
     catsMtl,
     googleContainer,
     kubernetesClient,
-    googleContainerV1
+    googleContainerV1,
+    sealerate
   )
 
   val newrelicDependencies = Seq(